Extend range for factorial tests and introduce double factorial
Extended the range of tested integers for factorial function from 0-20 up to 0-170 and introduced factorial function for doubles. The change will allow to better test the factorial function especially for large numbers. Added a test to check the factorialDouble function from the `CombinatoricsUtils` class, ensuring its correct implementation for the specific range of integers. This will help in capturing firstly, the large range of inputs and secondly, the behavior of the function when dealing with doubles.